// src/app/api/model/address-dto-in.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddressDtoIn { 
    additionalInfo?: string;
    /**
     * Type of address
     */
    addressType: AddressDtoInAddressTypeEnum;
    city?: string;
    country?: string;
    id?: number;
    partnershipOpportunityId?: number;
    postalCode?: string;
    primary?: boolean;
    state?: string;
    street?: string;
    userId?: number;
}
export enum AddressDtoInAddressTypeEnum {
    MAIN = 'MAIN',
    SECONDARY = 'SECONDARY',
    BILLING = 'BILLING',
    SHIPPING = 'SHIPPING',
    TEMPORARY = 'TEMPORARY'
};