// src/app/api/model/file-upload-request.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FileUploadRequest { 
    altText?: string;
    contentType?: FileUploadRequestContentTypeEnum;
    description?: string;
    fileExtension?: string;
    fileSize: number;
    filename?: string;
    uploadType?: string;
}
export enum FileUploadRequestContentTypeEnum {
    JPEG = 'image/jpeg',
    JPG = 'image/jpg',
    PNG = 'image/png',
    GIF = 'image/gif',
    WEBP = 'image/webp'
};